refactor: CMD-176 rename search to es_search
